Make metric names compliant with OpenMetrics conventions
OpenMetrics does not allow dots in metric names. [1]
For some reason the previous versions of Datadog agent worked even if
dots were in the metric name. Newer versions of Datadog agent do not
allow dots in OpenMetrics metric names and due to that this metric
was not reported to Datadog.

To fix this issue we now use metric name that is compliant with
OpenMetrics naming.

Example has been modified to show how to map metric names
between OpenMetrics and Datadog.

[1]: https://prometheus.io/docs/practices/naming/